feat: server action
