+++ .env.testnet.local
@@ -34,9 +34,6 @@ NEXT_PUBLIC_BUILDER_REGISTRY_ADDRESS=0xDf1ED223f6EC3cC5e57ed5BeaE8031d6BD37516d
 NEXT_PUBLIC_NFT_BOOSTER_DATA_URL="https://raw.githubusercontent.com/RootstockCollective/dao-frontend/boost_data/testnet.local/nft_boost_data"
 NEXT_PUBLIC_CR_MIGRATING=true
 
-# The graph
-NEXT_PUBLIC_THE_GRAPH_URL=https://api.studio.thegraph.com/query/98603/cr-poc/version/latest
-
 # Rif Wallet Services
 NEXT_PUBLIC_RIF_WALLET_SERVICES=http://localhost:3000/cors_bypass
 NEXT_PUBLIC_EVENTS_FROM_BLOCK=5784028
@@ -45,3 +42,8 @@ NEXT_PUBLIC_EVENTS_FROM_BLOCK=5784028
 NEXT_PUBLIC_PROXY_DESTINATION=https://dev.rws.app.rootstockcollective.xyz
 NEXT_PUBLIC_ENABLE_CORS_BYPASS=true
 NEXT_PUBLIC_ENABLE_FEATURE_V2_REWARDS=true
+
+# The graph
+THE_GRAPH_URL=https://api.studio.thegraph.com/query
+THE_GRAPH_API_KEY=98603
+THE_GRAPH_ID=cr-poc/v0.0.81
