+++ src/app/collective-rewards/rewards/backers/hooks/useGetBackerStakingHistory.ts
@@ -1,13 +1,7 @@
-import { AVERAGE_BLOCKTIME, THE_GRAPH_URL } from '@/lib/constants'
-import axios from 'axios'
+import { AVERAGE_BLOCKTIME } from '@/lib/constants'
 import { useQuery } from '@tanstack/react-query'
 import { Address } from 'viem'
-
-type Response = {
-  data: {
-    backerStakingHistory: BackerStakingHistory
-  }
-}
+import { fetchBackerStakingHistory } from '@/app/collective-rewards/rewards'
 
 export type BackerStakingHistory = {
   id: Address
@@ -24,40 +18,9 @@ export type GaugeStakingHistory = {
   lastBlockTimestamp_: string
 }
 
-const query = `
-    query($backer: Bytes){
-        backerStakingHistory(id: $backer) {
-            id
-            backerTotalAllocation_
-            accumulatedTime_
-            lastBlockTimestamp_
-            gauges_ {
-                gauge_
-                accumulatedAllocationsTime_
-                allocation_
-                lastBlockTimestamp_
-            }
-
-        }
-    }
-`
-
 export const useGetBackerStakingHistory = (backer: Address) => {
   const { data, isLoading, error } = useQuery({
-    queryFn: async () => {
-      const {
-        data: {
-          data: { backerStakingHistory },
-        },
-      } = await axios.post<Response>(THE_GRAPH_URL, {
-        query,
-        variables: {
-          backer,
-        },
-      })
-
-      return backerStakingHistory
-    },
+    queryFn: () => fetchBackerStakingHistory(backer),
     queryKey: ['backerStakingHistory', backer],
     refetchInterval: AVERAGE_BLOCKTIME,
   })
