+++ src/app/collective-rewards/rewards/backers/index.ts
@@ -1,6 +1,7 @@
 export * from './ABIBackers'
 export * from './context'
 export * from './hooks'
+export * from './actions'
 export * from './RBI'
 import { AllTimeShare } from './AllTimeShare'
 import { BackerRewardsTable } from './BackerRewardsTable'
