+++ src/lib/constants.ts
@@ -30,7 +30,6 @@ export const BUILDER_REGISTRY_ADDRESS = process.env.NEXT_PUBLIC_BUILDER_REGISTRY
 export const REWARD_DISTRIBUTOR_ADDRESS = process.env.NEXT_PUBLIC_REWARD_DISTRIBUTOR_ADDRESS as Address
 export const CR_MIGRATING = process.env.NEXT_PUBLIC_CR_MIGRATING === 'true'
 export const NFT_BOOSTER_DATA_URL = (process.env.NEXT_PUBLIC_NFT_BOOSTER_DATA_URL as string) ?? ''
-export const THE_GRAPH_URL = (process.env.NEXT_PUBLIC_THE_GRAPH_URL as string) ?? ''
 
 export const ADDRESS_ANIMATION_DURATION = 800
 export const AVERAGE_BLOCKTIME = 60_000
